+++ src/main/scala/com/raquo/laminar/keys/ReactiveEventProp.scala
@@ -1,14 +1,6 @@
 package com.raquo.laminar.keys
 
 import com.raquo.domtypes.generic.keys.EventProp
-import com.raquo.laminar.emitter.EventPropTransformation
-import com.raquo.laminar.nodes.ReactiveElement
 import org.scalajs.dom
 
-class ReactiveEventProp[Ev <: dom.Event](override val name: String) extends EventProp[Ev](name) {
-
-  def config[El <: ReactiveElement.Base](useCapture: Boolean = false): EventPropTransformation[Ev, Ev] = {
-    new EventPropTransformation(this, useCapture, processor = Some(_))
-  }
-
-}
+class ReactiveEventProp[Ev <: dom.Event](override val name: String) extends EventProp[Ev](name)