+++ src/main/scala/com/raquo/laminar/receivers/FocusReceiver.scala
@@ -1,12 +1,12 @@
 package com.raquo.laminar.receivers
 
 import com.raquo.airstream.eventstream.EventStream
-import com.raquo.laminar.modifiers.{FocusSetter, Setter}
+import com.raquo.laminar.modifiers.{Binder, FocusBinder}
 import com.raquo.laminar.nodes.ReactiveHtmlElement
 
 object FocusReceiver {
 
-  @inline def <--($isFocused: EventStream[Boolean]): Setter[ReactiveHtmlElement.Base] = {
-    FocusSetter($isFocused)
+  @inline def <--($isFocused: EventStream[Boolean]): Binder[ReactiveHtmlElement.Base] = {
+    FocusBinder($isFocused)
   }
 }